Remove task_categories foreign key from tasks
This created a dependency loop when trying to delete a user, since tasks
and task_categories both have foreign key constraints on user, so we
can't also have foreign key constraints between tasks and
task_categories. This makes it impossible to delete a user's tasks or
task_categories atomically during the user deletion.