Get login/logout process working
There is no CSS yet, but this allows a user to login and logout. The
user must already be defined in the db, e.g, done via the rails console.
This is a first step in bootstrapping some of the original login system
while making use of the new Rails 8 authentication system.